requirements.txt
# Core Dependencies
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
pandas==2.3.1
openpyxl==3.1.5
numpy==2.3.2

# QR Code Generation
qrcode==7.4.2
Pillow==10.1.0

# Web Server
eventlet==0.33.3
gunicorn==21.2.0

# Utilities
python-dateutil==2.9.0
pytz==2025.2
requests==2.31.0
colorama==0.4.6

# Development (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Windows specific (if needed)
# pywin32==306